♻️ Reopen: Issue/chen (#11)
* 🐛Fix: chen model nans (#8)

* 🐛Fix: #8
The torch.set_float32_matmul_precision was causing issues for large tensors matmuls on the cpu.
Removed this which should fix the bug. Due to the smaller batch size the other models were (probably) not affected.
But the experiments should be rerun nonetheless since we are using a higher precision now.