✨Add experiments for FID (#7)
* 🪄Reopen PR: sync feat/fid with dev

* ✨Add initial autoencoder

* ✨Add gradient loss

* ♻️ Reopen: Issue/chen (#11)

* 🐛Fix: chen model nans (#8)

* 🐛Fix: #8
The torch.set_float32_matmul_precision was causing issues for large tensors matmuls on the cpu.
Removed this which should fix the bug. Due to the smaller batch size the other models were (probably) not affected.
But the experiments should be rerun nonetheless since we are using a higher precision now.

* ✨Add Autoencoder

* ✨Add sweep for the autoencoder

* 🔧Refactor: Works with multiple leadtimes

* ✨Add FID computation

* 🔧Fix Test Dataloader handling

* 🐛Fix evaluation data selection
Rewrite dataloader to be faster and less error prone
Improvements to evaluation scrips -> move commonly used functions to utils
rename zarr.py -> zarr_dataset.py as this caused issues with the zarr module
Also configs have changed a bit due to teh new dataloader

* ✨Add polars and sns

* ✨Add plot for scores

* ✨Add Notebook for FID experiments

* 🌈Add HvPlot

* ✨Add saving of ae latents

* ✨Add Torchvision

* ✨Add ClassifierEncoder

* ✨Add visualization of FID

* ✨Add 1 more layer

* ✨Add sweeps

* ♻️ Upgrade Xarray

* ✨Add better evaluation saving

* ✨Better Evaluation

* 🪄Add standard split

* ♻️Refactor FID eval script

* 🧹Cleanup